Adds a table-oriented Diff generator class, and a HTML formatter for it to use for default output
