+++ tests/fixture_files/crew.csv
@@ -1,3 +1,3 @@
 First name,Last name,crew id, manager id, birth date, start date, end date, status,rank, pay rate, pay per
 James,Kirk, 1,,3/22/2233,8/8/2245,,ACTIVE,Captain,180000,year,
-Leonard, McCoy,2,1,1/1/2233,8/8/2245,,ACTIVE,Doctor,210000,year
+Leonard, McCoy,2,1,1/1/2233,8/8/2245,,ACTIVE,Doctor,210000,year,
